Offer gas price in more precise units.

+++ internal/types/gas_price.go
@@ -3,8 +3,8 @@ package types
 
 // GasPrice represents an extended gas price estimator.
 type GasPrice struct {
-	Fast    int32 `json:"fast"`
-	Fastest int32 `json:"fastest"`
-	SafeLow int32 `json:"safeLow"`
-	Average int32 `json:"average"`
+	Fast    float64 `json:"fast"`
+	Fastest float64 `json:"fastest"`
+	SafeLow float64 `json:"safeLow"`
+	Average float64 `json:"average"`
 }
